Upload image \<Project Description> Recognize faces in images or photos and automatically
\<Purpose>
Due to the era of SNS, which takes, posts, and shares photos and videos regardless of time and place, the number of cases of portrait rights infringement is increasing, so we developed an open source sw to prevent it.

\<Package and Version>
*opencv-python*
*numpy*

\<Test Example>
![Example Image](https://github.com/minxxeo/TermProject/commit/0d2ed04d4edefacec13a0df09ec6f1968cffb6e2#diff-3f8622c2567713163e9d2cff2cc06b47407364ffeadcb74ee51d34b7c768b4f8)

\<Running method>
It recognizes the face and eyes by using 'haarcascade_frontalface_default.xml' and 'haarcascade_eye.xml'.
Redefine the image through resize and mosaic the face.

\<Reference data\>
https://ponyozzang.tistory.com/599
https://github.com/devGyu97/opencv-face-recognition/blob/main/test.py